+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * static getter/setter should remain static. [#51](https://github.com/nojaf/telplin/issues/51)
 * Member with abstract decl and default impl uses "member" instead of "override" for impl signature entry [#53](https://github.com/nojaf/telplin/issues/53)
 * Don't add [<Class>] attribute for interface. [#55](https://github.com/nojaf/telplin/issues/55)
+* Access modifier from member is missing. [#56](https://github.com/nojaf/telplin/issues/56)
 
 ## 0.5.0 - 02-05-23